Allow view taxes with product manage permission
Update test snapshots

Update test snapshots
