+++ src/configuration/utils.ts
@@ -1,7 +1,9 @@
-import { PermissionEnum } from "@saleor/graphql";
+import { hasAnyPermissions } from "@saleor/auth/misc";
+import { PermissionEnum, UserFragment } from "@saleor/graphql";
 import { IntlShape } from "react-intl";
 
 import { createConfigurationMenu } from ".";
+import { MenuItem } from "./types";
 
 export const getConfigMenuItemsPermissions = (
   intl: IntlShape,
@@ -15,3 +17,9 @@ export const getConfigMenuItemsPermissions = (
       [],
     )
     .flat();
+
+export const hasUserMenuItemPermissions = (
+  menuItem: MenuItem,
+  user: UserFragment,
+): boolean =>
+  menuItem.permissions ? hasAnyPermissions(menuItem.permissions, user) : true;
